refactor(estructura): reorganizar carpetas dividiendo por páginas, componentes , servicios, etc
