+++ src/app/pages/registro/registro.component.html
