+++ src/app/pages/registro/registro.component.ts
